requirements.txt
python-dotenv
openai
inquirer